read bondcl only when using continue, not for zero
When using `zero` for txstart the bondcl properties are set by the
routine Particle according the the input of `nmlParticle`. If there were
some unwanted particle settings in the cnf file from a previous
simulation these were overwriting the settings generated from the
`nmlParticle` in the input file. This is changed, so that the particle
data is generated from the input file in case of `txstart` = `zero`

increment version