fix inclusion of fftw path and compiler selection
Before the subdirectory of the directory of the lib file was included.
The configure file was changed to look directly in the lib directory.
Additionally the setting of the fftwlib in the case of generating it
from the default settings was fixed in the makefile.

The selection of the compiler now also takes into account what compiler
is available for the parallel compilation.