fixing cell list in the case on single cell case
When the box length is so small, that the systems fits into one cell in
the direction of a specific dimension, whe do not use the celllist, but
rather use a single cell.

It was fixed that the parameter cellsize and cellsizei are also set
correctly when the celllist is initialized so small that it is not
needed. This way pcellro works correctly.